# Reduce bytecode size of PlatformDependent0.equals.