@normanmaurer 2016-09-07
@buchgr please fix test error ;)