@normanmaurer 2016-09-09
Cherry-picked as 67d3a78123fa3faa85c1a150bd4ee69425079b3d into 4.1