@Scottmitch 2016-09-10
@buchgr - sorry for the regression and thanks for the fix ... i thought I had benchmarked this change but perhaps the difference was in the noise.