@buchgr 2016-09-10
@Scottmitch 

As you can see, in the JMH benchmarks the numbers are pretty much the same. In my application it does make a considerable difference. I assume it is due to some optimizations the JIT compiler is able to do, when the equals method can be inlined with the other code. So I think this is something a microbenchmark cannot show.