@doom369 2016-09-10
@buchgr you need add to benchmarks wrapping code that invokes `equals`. And you can manage inline options from benchmark.