@buchgr 2016-09-11
@doom369 

the method is small enough to be inlined. so I think that's not necessary now.