@Scottmitch 2016-09-14
@buchgr +1 for real world tests.

Were you able to verify that before your change this method (and/or other in the call stack) was not inlined, and after they were inlined?